--- src/metrics/IDataListener.java
@@ -1,6 +1,0 @@
-package metrics;
-
-public interface IDataListener
-{
-    void getNewData(String name, Object value, Double timestamp);
-}
