+++ src/metrics/IDataReader.java
@@ -2,6 +2,5 @@ package metrics;
 
 public interface IDataReader
 {
-    void addNewDataListener(IDataListener newDataListener);
     void startReading();
 }
