--- src/metrics/IDataSaver.java
@@ -1,6 +1,0 @@
-package metrics;
-
-public interface IDataSaver
-{
-    void addEntry(String name, Object value, Double timestamp);
-}
