+++ src/metrics/IDataStreamer.java
@@ -3,6 +3,5 @@ package metrics;
 public interface IDataStreamer
 {
     void addStreamListener(IStreamListener streamListener);
-
     void startStreaming();
 }
