--- src/metrics/NoLogging.java
@@ -1,9 +1,0 @@
-package metrics;
-
-public class NoLogging implements IDataSaver
-{
-    public void addEntry(String name, Object value, Double timestamp)
-    {
-        // No action
-    }
-}
